requirements.txt
Flask
pycryptodome
flask-cors 